Agregando archivos par perfil

+++ src/main/java/com/example/edelec/repositories/UsuarioRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
     @Query("select u FROM  Usuario u WHERE u.nombre=:nombre")
     List<Usuario> buscarnombredeUsuario(@Param("nombre") String nombre);
